+++ src/modules/data/group/group-classes.composable.ts
@@ -1,15 +1,13 @@
-import { ClassInfo } from "./types/class-info";
-import { Pagination } from "./types/commons";
-import { SortOrder } from "./types/sort-order.enum";
 import { useSafeAxiosTask } from "@/composables/async-tasks.composable";
 import { useI18nGlobal } from "@/plugins/i18n";
+import { Pagination } from "@/store/types/commons";
+import { SortOrder } from "@/store/types/sort-order.enum";
 import { $axios } from "@/utils/api";
 import { ClassSortQueryType, GroupApiFactory, SchoolYearQueryType } from "@api-server";
-import { GroupMapper } from "@data-group";
+import { ClassInfo, GroupMapper } from "@data-group";
 import { ref } from "vue";
 
-// TODO: where to move this file? modules/data/??  group? administration? classes?
-export const useGroup = () => {
+export const useGroupClasses = () => {
 	const { t } = useI18nGlobal();
 	const groupApi = GroupApiFactory(undefined, "/v3", $axios);
