+++ src/modules/data/group/index.ts
@@ -1,3 +1,4 @@
+export * from "./group-classes.composable";
 export * from "./GroupApi.composable";
 export * from "./groupListState.composable";
 export * from "./GroupMapper";
