+++ src/modules/data/group/types/class-info.ts
