+++ src/pages/administration/ClassOverview.page.unit.ts
@@ -1,12 +1,12 @@
 import ClassOverview from "./ClassOverview.page.vue";
+import { ClassRootType } from "@/modules/data/group/types/class-info";
 import GroupModule from "@/store/group";
 import SchoolsModule from "@/store/schools";
-import { ClassRootType } from "@/store/types/class-info";
 import { Pagination } from "@/store/types/commons";
 import { School, Year } from "@/store/types/schools";
 import { SortOrder } from "@/store/types/sort-order.enum";
 import * as confirmDialogUtils from "@/utils/confirmation-dialog.utils";
-import { GROUP_MODULE_KEY, SCHOOLS_MODULE_KEY } from "@/utils/inject";
+import { SCHOOLS_MODULE_KEY } from "@/utils/inject";
 import {
 	classInfoFactory,
 	courseFactory,
